Remove batch limit in training cc2ftr for JIT Defect Prediction
There seems to be a bug in this file, where the training batches are limited to only the first 10 batches. This means the model is only trained on a very small subset of the dataset. The fact that no such batch limit occurs in the other two applications of CC2Vec, leads me to believe this is unintended.